The Office Text Analysis - Full Quote Databases
Restructured the filesystem into data and scripts. Processing was used
to parse the large (4MB) quotes.json file into a CSV file, the CSV was
then uploaded to a MySQL Database and then downloaded and included in
the data folder.

There are many spelling and text encoding errors that still need to be
fixed, but for the most part, the data is ready to be analyzed.